Added new Pojos and Not Founds

+++ src/main/java/com/seachangesimulations/scorp/controller/ScorpRestCrudController.java
@@ -19,23 +19,25 @@ import com.seachangesimulations.scorp.service.ObjectService;
 
 /**
  * Handles the CRUD operations for base domain objects.
- *
+ * No error checking. For example if a foreign key does 
+ * not exist.
  */
 @RestController
-@RequestMapping("/rest")
+@RequestMapping("/rest")  // scorp/rest if run from Tomcat, just /rest if RunAs->SpringBootApp
 public class ScorpRestCrudController {
 	
-	@Autowired
+	@Autowired  // Dependency Inject this value.
 	protected ObjectService objectService;
 	
-	ScorpRestCrudController(){
-		
+	ScorpRestCrudController() {		
 	}
 	
+	// ------ Methods in CRUD order ------ 
 	/**
 	 * Create.
+	 * If object already exists ...
 	 * 
-	 * @param objectName
+	 * @param objectName - a String such as actor or phase representing the object to create.
 	 * @param linkedHashMap
 	 * @return
 	 */
@@ -49,7 +51,7 @@ public class ScorpRestCrudController {
 	/**
 	 * Read All.
 	 * 
-	 * @param objectName
+	 * Get all Objects - Read All. For example rest/actor returns all actors.
 	 * @return
 	 */
 	@SuppressWarnings("unchecked")
@@ -71,14 +73,27 @@ public class ScorpRestCrudController {
 	@RequestMapping(value="/{objectClass}/{objectId}", method = RequestMethod.GET, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity findOne(@PathVariable String objectClass, @PathVariable Long objectId) {
 		Object object = this.objectService.findById(objectClass, objectId);
+		// 6.20.18 MJS - return 404 Not Found if object not in database
+		if (object == null) {
+			return new ResponseEntity(HttpStatus.NOT_FOUND);
+		}
 		return new ResponseEntity(object, HttpStatus.OK);
 	}
 	
-	/* UPDATE */
-	
+	/**
+	 * Update
+	 * 
+	 * @param objectClass
+	 * @param objectId
+	 * @return
+	 */
 	@RequestMapping(value="/{objectClass}/{objectId}", method = RequestMethod.PUT, produces = MediaType.APPLICATION_JSON_VALUE)
 	public ResponseEntity update(@PathVariable String objectClass, @PathVariable Long objectId, @RequestBody LinkedHashMap linkedHashMap) {
-
+		Object object = this.objectService.findById(objectClass, objectId);
+		// 6.20.18 MJS - return 404 Not Found if object not in database
+		if (object == null) {
+			return new ResponseEntity(HttpStatus.NOT_FOUND);
+		}
 		this.objectService.update(objectClass, objectId, linkedHashMap);
 		return new ResponseEntity(linkedHashMap, HttpStatus.OK);
 	}
@@ -93,7 +108,12 @@ public class ScorpRestCrudController {
 	@SuppressWarnings("rawtypes")
 	@RequestMapping(value="/{objectClass}/{objectId}", method = RequestMethod.DELETE, produces = MediaType.APPLICATION_JSON_VALUE)
 	public @ResponseBody HttpEntity delete(@PathVariable String objectClass, @PathVariable Long objectId) {
+		Object object = this.objectService.findById(objectClass, objectId);
+		// 6.20.18 MJS - return 404 Not Found if object not in database
+		if (object == null) {
+			return new ResponseEntity(HttpStatus.NOT_FOUND);  // 404
+		}
 		this.objectService.delete(objectClass, objectId);
-		return new ResponseEntity(HttpStatus.NO_CONTENT);
+		return new ResponseEntity(HttpStatus.NO_CONTENT);   // 204
 	}
 }
